KTLN-308 : Sealed Class vs Enum in Kotlin
