+++ core-kotlin-modules/core-kotlin-lang-oop-2/src/test/kotlin/com/baeldung/sealedclasses/vs/enums/OsSealedUnitTest.kt
@@ -1,4 +1,4 @@
-package com.baeldung.kotlin.enums.vs.sealed.classes
+package com.baeldung.sealedclasses.vs.enums
 
 import org.junit.jupiter.api.Test
 import kotlin.test.assertEquals
@@ -13,11 +13,21 @@ class OsSealedUnitTest {
     }
 
     private fun detectOsByUsingWhen(osSealed: OsSealed) {
+        assertEquals("Called from parent sealed class", osSealed.getTextParent())
         when (osSealed) {
             is OsSealed.Linux -> assertEquals("Linux by Open-Source - value=1", osSealed.getText(1))
             is OsSealed.Mac -> assertEquals("Mac by Apple - released at 2001", osSealed.doSomething())
             is OsSealed.Windows -> assertEquals(5, osSealed.getNumber("Text!"))
             else -> assert(osSealed is OsSealed.Unknown)
+        }
+    }
+
+    private fun detectOsByUsingWhenPrint(osSealed: OsSealed) {
+        when (osSealed) {
+            OsSealed.Linux -> println("${osSealed.company} - Linux Operating System")
+            OsSealed.Mac -> println("${osSealed.company} - Mac Operating System")
+            OsSealed.Windows -> println("${osSealed.company} - Windows Operating System")
+            else -> println(osSealed.company)
         }
     }
 }
